# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SaaS portfolio part2",
        "version": "0.1"
    },
    "paths": {
        "/api/register": {
            "post": {
                "summary": "Register new user",
                "operationId": "72310a6d9260c39f210ba54c35a12b41",
                "parameters": [
                    {
                        "name": "given_name",
                        "in": "query",
                        "description": "given name of the user",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "family_name",
                        "in": "query",
                        "description": "family name of the user",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email address of the user",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password of the user",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully Register new User"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "summary": "login user",
                "operationId": "a36859cc9285d70ef9eb697ebd3b2733",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email address of the user",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "password of the user",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully logged in"
                    },
                    "401": {
                        "description": "Login information is invalid."
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "summary": "logout user - Needs Futher testing",
                "operationId": "b244ee3d05d3f6a167e0bf9e2a322e07",
                "parameters": [
                    {
                        "name": "access_token",
                        "in": "header",
                        "description": "email address of the user",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully logged out the User"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/authors": {
            "get": {
                "summary": "Finds the list of all authors ",
                "operationId": "76d3ed20fa87acaf0c03065b03ec5cfa",
                "responses": {
                    "200": {
                        "description": "Browse the list of all authors"
                    }
                }
            },
            "post": {
                "summary": "Store new author",
                "operationId": "78b0dc6c2dd4a8e87bda6852cefa9d9d",
                "parameters": [
                    {
                        "name": "given_name",
                        "in": "query",
                        "description": "given name of the author",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "family_name",
                        "in": "query",
                        "description": "family name of the book",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully Store new author"
                    }
                }
            }
        },
        "/api/authors/{id}": {
            "get": {
                "summary": "Finds authors by author id",
                "operationId": "3015af83cc37c3dcfb442fa4c7a53565",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Author id to filter by",
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show selected Author"
                    },
                    "404": {
                        "description": "Show error when Author not Found"
                    }
                }
            },
            "delete": {
                "summary": "Delete the author",
                "operationId": "9e2cc8b9f0b308755b27603b37d2b688",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Author id to filter by",
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully delete the author"
                    },
                    "404": {
                        "description": "Show error when author not Found"
                    }
                }
            },
            "patch": {
                "summary": "Update the author",
                "operationId": "39fa31c35696bb859a5b54c2e13792d1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the author",
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "given_name",
                        "in": "query",
                        "description": "given name of the author",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "family_name",
                        "in": "query",
                        "description": "family name of the book",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update the author"
                    },
                    "404": {
                        "description": "Show error when author not Found"
                    }
                }
            }
        },
...
    }
}